% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_categories.R
\name{broad_cn}
\alias{broad_cn}
\title{something}
\usage{
broad_cn(seg = seg, log_ratio_threshold = 0.2)
}
\arguments{
\item{log_ratio_threshold}{}
}
\value{
binary table of broad copy number alterations in the 341 IMPACT gene panel
}
\description{
Broad copy number alterations
}
