% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_categories.R
\name{cn_burden}
\alias{cn_burden}
\title{something}
\usage{
cn_burden(seg = seg, log_ratio_threshold = 0.2)
}
\arguments{
\item{log_ratio_threshold}{}
}
\value{
copy number burden of samples from selected cancer types
}
\description{
Copy number burden
}
