% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_categories.R
\name{focal_cn_portal}
\alias{focal_cn_portal}
\title{something}
\usage{
focal_cn_portal(cna = cn)
}
\arguments{
\item{cna}{}
}
\value{
binary table of focal copy number alterations in the 341 IMPACT gene panel
}
\description{
Focal amplifications and homozygous deletions portal style - MSK-IMPACT 341

This function creates a feature table of focal copy number alterations of type:
Amp (amplification, cn == 2)
HomDel (homozygous deletion, cn == -1)
}
