% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_categories.R
\name{fusions}
\alias{fusions}
\title{something}
\usage{
fusions(SV = SV)
}
\arguments{
\item{SV}{}
}
\value{
binary table of structural variants involving genes in the 341 IMPACT gene panel
}
\description{
Fusions
}
