% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GDDmetrics.R
\name{GDDmetrics}
\alias{GDDmetrics}
\title{GDD metrics}
\usage{
GDDmetrics(data, lev = NULL, model = NULL)
}
\arguments{
\item{data}{input data}

\item{lev}{input leve}

\item{model}{input model}
}
\value{
stats
}
\description{
GDD metrics
}
