% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_feature_table.R
\name{gen_feature_table}
\alias{gen_feature_table}
\title{gen_feature_table}
\usage{
gen_feature_table(repository_folder)
}
\arguments{
\item{repository_folder}{}
}
\value{
feature_table
}
\description{
gen_feature_table
}
