% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_feature_table.R
\name{generate_feature_table}
\alias{generate_feature_table}
\title{generate_feature_table}
\usage{
generate_feature_table(arg_line = NA)
}
\arguments{
\item{repository_folder}{}
}
\value{
feature_table
}
\description{
generate_feature_table
}
