% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_new.R
\name{get_fc}
\alias{get_fc}
\title{get_fc}
\usage{
get_fc(rfe_model, feature_table, mClass = NULL)
}
\arguments{
\item{rfe_model}{train_class}

\item{feature_table}{gen_fature_table}

\item{mClass}{NULL}
}
\value{

}
\description{
get_fc
}
