% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_new.R
\name{get_xval_predictions}
\alias{get_xval_predictions}
\title{get_xval_predictions}
\usage{
get_xval_predictions(rfe_model, N_top_cancer_types = 3)
}
\arguments{
\item{rfe_model}{train_class}

\item{N_top_cancer_types}{3}
}
\value{

}
\description{
get_xval_predictions
}
