% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_categories.R
\docType{data}
\name{hotspot_list}
\alias{hotspot_list}
\title{Hotspots - MSK-IMPACT 341}
\format{An object of class \code{character} of length 3746.}
\usage{
hotspot_list
}
\arguments{
\item{maf}{}
}
\value{
binary table of hotpots in the 341 IMPACT gene panel
}
\description{
Hotspots - MSK-IMPACT 341
}
\keyword{datasets}
