% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_categories.R
\name{hotspots}
\alias{hotspots}
\title{something}
\usage{
hotspots(maf = maf)
}
\arguments{
\item{maf}{}
}
\value{
binary table of hotpots in the 341 IMPACT gene panel
}
\description{
Hotspots - MSK-IMPACT 341
}
