% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_categories.R
\name{mutations}
\alias{mutations}
\title{something}
\usage{
mutations(maf = maf)
}
\arguments{
\item{maf}{}
}
\value{
binary table of mutations in the 341 IMPACT gene panel
}
\description{
Mutations - MSK-IMPACT 341
}
