% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_new.R
\name{pred_new}
\alias{pred_new}
\title{pred_new}
\usage{
pred_new(
  rfe_model,
  feature_table,
  model_function_list,
  N_top_cancer_types = 3,
  all_cancer_types = TRUE,
  N_top_variables = 10,
  all_variables = FALSE
)
}
\arguments{
\item{rfe_model}{train_class}

\item{feature_table}{gen_fature_table}

\item{model_function_list}{rfcal}

\item{N_top_cancer_types}{3}

\item{all_cancer_types}{NULL}

\item{N_top_variables}{10}

\item{all_variables}{FALSE}
}
\value{

}
\description{
pred_new
}
