% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_new.R
\name{predict_table}
\alias{predict_table}
\title{predict_table}
\usage{
predict_table(
  rfe_model,
  feature_table,
  model_function_list,
  N_top_cancer_types = 3
)
}
\arguments{
\item{rfe_model}{train_class}

\item{feature_table}{gen_fature_table}

\item{model_function_list}{rfcal}

\item{N_top_cancer_types}{3}
}
\value{

}
\description{
predict_table
}
