% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_categories.R
\name{truncating_mutations}
\alias{truncating_mutations}
\title{something}
\usage{
truncating_mutations(maf = maf)
}
\arguments{
\item{maf}{}
}
\value{
binary table of truncating mutations in the 341 IMPACT gene panel
}
\description{
Truncating mutations - MSK-IMPACT341
}
