% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_categories.R
\name{mutation_count}
\alias{mutation_count}
\title{something}
\usage{
mutation_count(maf)
}
\arguments{
\item{maf}{MAF format file}
}
\value{
table of mutation counts in the 341 IMPACT gene panel
}
\description{
Mutation count
}
